requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
fastapi==0.92.0
h11==0.14.0
idna==3.4
Pillow==9.4.0
pipreqs==0.4.11
pydantic==1.10.5
python-multipart==0.0.5
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
yarg==0.1.9
